Release: Add npx to bower command
This isn't expected to be globally installed in an isolated/secure
dev container with only Node.js and npm pre-installed. Rather than
installing it ad-hoc in /tmp or globally outside the container,
prescribe npx instead which gets the best of both ad-hoc and still
possible to do within the isolated environment.